UI code refactor
